requirements.txt
torch
torchvision
timm
pillow
PyYAML
